Get rid of the workspace directory, create a folder on the desktop
This is a big change so we're gonna have to go over it with TAs
beforehand. The shared folder location on the host (i.e. student laptop)
will be:

    ~/Desktop/railsbridge

This is the location of the Desktop in both Windows 7+ and OS X.
"Desktop" is capitalized. "railsbridge" should be lowercase always in
case someone's computer has a case-sensitive filesystem. On the guest
VM, we will now use the default share location of

    /vagrant

This changes some command line extra-credit stuff to match, so that all
the commands are still valid, and so those could use some improvement to
make the explanations flow more naturally again. We'll figure it out at
the workshop.