Fix the double flash message
The rails scaffold adds a line to the index and show view of every
scaffold:

```
<p id="notice"><%= notice %></p>
```

When we ask students to add flash messages to the layout, the additional
scaffold-added notice messages are also displayed. A lot of people
reported this at the last workshop.

sites/en/intro-to-rails/redirect_to_the_topics_list_after_creating_a_new
_topic.step