Refactor console/irb commands for consistency
Recently, the irb method was added. This displays data that the student
can type into IRB, similar to the terminal AKA the console. The console
methods had `console_with_` and `console_without_message` versions that
also allowed a caption to show up with the code.

This adds the `_with_` and `_without_message` versions of the method
for irb also. This will allow irb instructions to have all the same
features as terminal/console instructions.