Update steps for "redirect after creating a new topic"
Rails 5 now adds a flash message for you, however it's only a notice.
It didn't make a ton of sense to instruct students to remove the flash
from the topics index and add a different flash to the application
layout. I also added some explanation at the end about where the flash
code can be found